Inline concerns
Part of refactoring to remove ECP.
These concerns are only used in these classes, inlining them to make
tracking the difference between ECP and LUP easier.
